Service layer, component registration

+++ src/main/java/br/com/cursonelio/nelio/repository/UserRepository.java
@@ -3,5 +3,7 @@ package br.com.cursonelio.nelio.repository;
 import br.com.cursonelio.nelio.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
 }
